% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{analyze_survival}
\alias{analyze_survival}
\title{Analyze Survival}
\usage{
analyze_survival(data)
}
\arguments{
\item{data}{Input data.frame from generate_survival_data()}
}
\value{
A survfit object
}
\description{
Perform survival analysis
}
\details{
Fits Kaplan-Meier survival curves by sex
}
