% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synthetic_lung}
\alias{synthetic_lung}
\title{Synthetic Lung Cancer Survival Data}
\format{
A data frame with 300 rows and 6 variables:
\describe{
\item{patient_id}{Patient ID (1-300)}
\item{time_days}{Survival time in days}
\item{time_years}{Survival time in years}
\item{status}{Event status (1=death, 0=censored)}
\item{sex}{Patient sex (Male/Female)}
\item{age}{Age at diagnosis}
}
}
\source{
Simulated data
}
\usage{
synthetic_lung
}
\description{
A simulated dataset containing survival times for 300 patients
}
\keyword{datasets}
