% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survival_analysis.R
\name{analyze_survival}
\alias{analyze_survival}
\title{Run survival analysis}
\usage{
analyze_survival(data, time_var = "time_days")
}
\arguments{
\item{data}{Input data}

\item{time_var}{Time variable (in days)}
}
\description{
Run survival analysis
}
